chore: move copy-prototype-methods up one level
