+++ lib/copy-prototype-methods.mjs
@@ -1,5 +1,4 @@
-import throwsOnProto from "../throws-on-proto.mjs";
-var call = Function.call;
+import throwsOnProto from "./throws-on-proto.mjs";
 
 var disallowedProperties = [
     // ignore size because it throws from Map
@@ -22,6 +21,8 @@ if (throwsOnProto) {
  * @returns    {object}  A new object that has the same methods of the prototype
  */
 function copyPrototypeMethods(prototype) {
+    const call = Function.call;
+
     // eslint-disable-next-line @sinonjs/no-prototype-methods/no-prototype-methods
     return Object.getOwnPropertyNames(prototype).reduce(function (
         result,
