+++ lib/copy-prototype-methods.test.mjs
