+++ lib/prototypes/string.mjs
@@ -1,3 +1,3 @@
-import copyPrototype from "./copy-prototype-methods.mjs";
+import copyPrototype from "../copy-prototype-methods.mjs";
 
 export default copyPrototype(String.prototype);
